requirements.txt
evernote3
requests
flask
python-dotenv
fastapi
uvicorn
keyring
pydantic
beautifulsoup4
markdownify
pytest
lxml 